requirements.txt
fastapi
uvicorn
Pillow
numpy
python-multipart
tensorflow-lite